--- sympy/relation/eqntools/__init__.py
@@ -1,9 +1,0 @@
-"""
-Module to implement equation manipulation algorithm.
-"""
-
-__all__ = ['rearrange', 'eqnsimp', 'solveeqn']
-
-from .rearrange import rearrange
-from .simplify import eqnsimp
-from .solveeqn import solveeqn
