+++ sympy/relation/reltools/__init__.py
@@ -1,7 +1,9 @@
 """
-Module to implement symbolic manipulation algorithm of relations.
+Module to implement equation manipulation algorithm.
 """
 
-__all__ = ['eqnsimp']
+__all__ = ['rearrange', 'eqnsimp', 'solveeqn']
 
+from .rearrange import rearrange
 from .simplify import eqnsimp
+from .solveeqn import solveeqn
