+++ sympy/relation/reltools/rearrange.py
