+++ sympy/relation/reltools/solveeqn.py
